Globalize PII object mocking.
